Flatten Event down to a final class, ArrayEvent
