+++ library/core/qsl-base/src/main/java/org/quiltmc/qsl/api/event/package-info.java
@@ -1,11 +1,7 @@
 /**
  * Quilt standard library event apis.
  *
- * This package contains base objects for implementing events. Actual events hooking into the game can be found in
- * different packages.
- *
- * @see org.quiltmc.qsl.api.event.Event
- * @see org.quiltmc.qsl.api.event.EventFactory
+ * @see org.quiltmc.qsl.api.event.ArrayEvent
  */
 
 package org.quiltmc.qsl.api.event;
