--- library/core/qsl-base/src/main/java/org/quiltmc/qsl/impl/event/EventFactoryImpl.java
@@ -1,14 +1,0 @@
-package org.quiltmc.qsl.impl.event;
-
-import java.util.function.Function;
-
-import org.quiltmc.qsl.api.event.Event;
-
-public final class EventFactoryImpl {
-	public static <T> Event<T> createArrayEvent(Class<? super T> type, Function<T[], T> implementation) {
-		return new ArrayEvent<>(type, implementation);
-	}
-
-	private EventFactoryImpl() {
-	}
-}
